requirements.txt
emoji==1.2.0
pandas==0.25.3
plotly==4.14.3
numpy==1.19.2
psaw==0.1.0
streamlit-nightly==0.79.1.dev20210329
SQLAlchemy==1.4.5
yfinance==0.1.55
tweepy==3.10.0
vaderSentiment==3.3.2
